IGNITE-14385: fix comments
